can add posts now

+++ src/main/java/kat/venusrestblog/controller/PostsController.java
@@ -1,61 +1,90 @@
 package kat.venusrestblog.controller;
+import kat.venusrestblog.data.Category;
 import kat.venusrestblog.data.User;
 import kat.venusrestblog.data.Post;
+import kat.venusrestblog.error.misc.FieldHelper;
+import kat.venusrestblog.repository.CategoriesRepository;
 import kat.venusrestblog.repository.PostsRepository;
 import kat.venusrestblog.repository.UsersRepository;
+import kat.venusrestblog.security.OAuthConfiguration;
+import kat.venusrestblog.service.EmailService;
+import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+@AllArgsConstructor
 @RestController
 @RequestMapping(value = "/api/posts", produces = "application/json")
-
 public class PostsController {
+    private EmailService emailService;
     private PostsRepository postsRepository;
     private UsersRepository usersRepository;
+    private CategoriesRepository categoriesRepository;
 
-    public PostsController(PostsRepository postsRepository) {
-        this.postsRepository = postsRepository;
+    @GetMapping("")
+    public List<Post> fetchPosts() {
+        return postsRepository.findAll();
     }
-@GetMapping (value = "")
-public List<Post> fetchPosts (){
-    return postsRepository.findAll();
-}
 
     @GetMapping("/{id}")
     public Optional<Post> fetchPostById(@PathVariable long id) {
-        return postsRepository.findById(id);
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
+        return optionalPost;
     }
 
     @PostMapping("")
-    public void createPost(@RequestBody Post newPost){
-        User author = usersRepository.findById(1L).get();
+    public void createPost(@RequestBody Post newPost, OAuth2Authentication auth) {
+        String userName = auth.getName();
+        User author = usersRepository.findByUserName(userName);
         newPost.setAuthor(author);
+
+        newPost.setCategories(new ArrayList<>());
+
+        // use first 2 categories for the post by default
+        Category cat1 = categoriesRepository.findById(1L).get();
+        Category cat2 = categoriesRepository.findById(2L).get();
+
+        newPost.getCategories().add(cat1);
+        newPost.getCategories().add(cat2);
+
         postsRepository.save(newPost);
+
+        emailService.prepareAndSend(newPost, "Hey man you made a post", "See subject");
     }
 
     @DeleteMapping("/{id}")
-    public void deletePostsById (@PathVariable long id){
+    public void deletePostById(@PathVariable long id) {
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
         postsRepository.deleteById(id);
     }
+
     @PutMapping("/{id}")
-    public void updatePost( @RequestBody Post updatedPost, @PathVariable long id){
+    public void updatePost(@RequestBody Post updatedPost, @PathVariable long id) {
+        Optional<Post> originalPost = postsRepository.findById(id);
+        if(originalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post " + id + " not found");
+        }
+
+        // in case id is not in the request body (i.e., updatedPost), set it
+        // with the path variable id
         updatedPost.setId(id);
-        postsRepository.save(updatedPost);
-////    find the post you want to update in the posts list
-//        Post post = findPostById(id);
-//        if(post == null){
-//            System.out.println("Post not found");
-//        }else{
-//            if (updatedPost.getTitle() != null){
-//                post.setTitle(updatedPost.getTitle());
-//            }
-//            if(updatedPost.getContent() != null){
-//                post.setContent(updatedPost.getContent());
-//            }
-//            return;
-//        }
-//        throw new RuntimeException("Post not found");
+
+        // copy any new field values FROM updatedPost TO originalPost
+        BeanUtils.copyProperties(updatedPost, originalPost.get(), FieldHelper.getNullPropertyNames(updatedPost));
+
+        postsRepository.save(originalPost.get());
     }
 }
