+++ src/main/java/kat/venusrestblog/controller/UsersController.java
@@ -1,15 +1,21 @@
 package kat.venusrestblog.controller;
 
 import kat.venusrestblog.data.User;
+import kat.venusrestblog.data.UserRole;
+import kat.venusrestblog.dto.UserFetchDTO;
+import kat.venusrestblog.error.misc.FieldHelper;
 import kat.venusrestblog.repository.UsersRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Size;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,15 +23,32 @@ import java.util.Optional;
 @RestController
 @RequestMapping(value = "/api/users", produces = "application/json")
 public class UsersController {
+    private PasswordEncoder passwordEncoder;
     private UsersRepository usersRepository;
     @GetMapping(value = "")
-    public List<User> fetchUsers (){
-        return usersRepository.findAll();
+    public List<UserFetchDTO> fetchUsers() {
+//        return usersRepository.fetchUserDTOs();
+        List<User> users = usersRepository.findAll();
+        List<UserFetchDTO> userDTOs = new ArrayList<>();
+
+        for(User user : users) {
+            UserFetchDTO userDTO = new UserFetchDTO();
+            userDTO.setId(user.getId());
+            userDTO.setUserName(user.getUserName());
+            userDTO.setEmail(user.getEmail());
+            userDTOs.add(userDTO);
+        }
+
+        return userDTOs;
     }
 
     @GetMapping("/{id}")
     public Optional<User> fetchUserById(@PathVariable long id) {
-        return usersRepository.findById(id);
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User" + id +"not found");
+        }
+        return optionalUser;
     }
     @GetMapping("/me")
     private Optional<User> fetchMe() {
@@ -49,36 +72,60 @@ public class UsersController {
 
 
     @PostMapping("/create")
-    public void createUser(@RequestBody User newUser){
+    public void createUser(@RequestBody User newUser) {
+        // TODO: validate new user fields
+        newUser.setRole(UserRole.USER);
+
+        String plainTextPassword = newUser.getPassword();
+        String encryptedPassword = passwordEncoder.encode(plainTextPassword);
+        newUser.setPassword(encryptedPassword);
+
+        // don't need the below line at this point but just for kicks
         newUser.setCreatedAt(LocalDate.now());
         usersRepository.save(newUser);
     }
 
+
     @DeleteMapping("/{id}")
     public void deleteUserById(@PathVariable long id) {
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if (optionalUser.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User" + id + "not found");
+        }
         usersRepository.deleteById(id);
     }
     @PutMapping("/{id}")
     public void updateUser( @RequestBody User updatedUser, @PathVariable long id) {
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+        User originalUser = optionalUser.get();
+        // merge the changed data in updatedUser with originalUser
+        BeanUtils.copyProperties(updatedUser, originalUser, FieldHelper.getNullPropertyNames(updatedUser));
+        // originalUser now has the merged data (changes + original data)
         updatedUser.setId(id);
         usersRepository.save(updatedUser);
     }
     @PutMapping("/{id}/updatePassword")
     private void updatePassword(@PathVariable Long id, @RequestParam(required = false) String oldPassword, @Valid @Size(min = 3) @RequestParam String newPassword){
-        User user = usersRepository.findById(id).get();
-//        if(user == null){
-//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User id" + id + "not found");
-//        }
-//        to compare the old password with saved password
-        if (!user.getPassword().equals(oldPassword)){
-
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"asmcrayy");
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
         }
-//        to validate new password
-        if (newPassword.length() < 3){
 
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "new password length must be at least 3 characters");
+        User user = optionalUser.get();
+
+        // compare old password with saved pw
+        if(!user.getPassword().equals(oldPassword)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "amscray");
         }
+
+        // validate new password
+        if(newPassword.length() < 3) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "new pw length must be at least 3 characters");
+        }
+
         user.setPassword(newPassword);
         usersRepository.save(user);
     }
